Conflicts:
	src/database/Genome.java

+++ src/database/subClasses/FileMethods.java
@@ -7,14 +7,14 @@ import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 
-import database.Experiment;
 import database.FilePathGenerator;
-import database.FileTuple;
+import database.containers.Experiment;
+import database.containers.FileTuple;
 
 /**
  * Class that contains all the methods for adding,changing, getting and removing
  * Files in the database. This class is a subClass of databaseAcessor.java.
- * 
+ *
  * date: 2014-05-14 version: 1.0
  */
 public class FileMethods {
@@ -23,20 +23,19 @@ public class FileMethods {
     private FilePathGenerator fpg;
     private ExperimentMethods expMethods;
 
-
     /**
      * Constructor for the fileMethod object.
-     * 
+     *
      * @param connection
      *            Connection, the connection to the database.
      */
     public FileMethods(Connection connection, FilePathGenerator filePG,
             ExperimentMethods expMethods) {
+
         conn = connection;
         fpg = filePG;
         this.expMethods = expMethods;
     }
-
 
     /**
      * @param expID
@@ -77,6 +76,7 @@ public class FileMethods {
             throws SQLException, IOException {
 
         Experiment e = expMethods.getExperiment(expID);
+
         if (e == null) {
             throw new IOException("The experiment " + expID
                     + " does not exist!");
@@ -109,18 +109,18 @@ public class FileMethods {
         stmt.setString(1, path);
 
         switch (fileType) {
-            case FileTuple.RAW:
-                stmt.setString(2, "Raw");
-                break;
-            case FileTuple.PROFILE:
-                stmt.setString(2, "Profile");
-                break;
-            case FileTuple.REGION:
-                stmt.setString(2, "Region");
-                break;
-            default:
-                stmt.setString(2, "Other");
-                break;
+        case FileTuple.RAW:
+            stmt.setString(2, "Raw");
+            break;
+        case FileTuple.PROFILE:
+            stmt.setString(2, "Profile");
+            break;
+        case FileTuple.REGION:
+            stmt.setString(2, "Region");
+            break;
+        default:
+            stmt.setString(2, "Other");
+            break;
         }
 
         stmt.setString(3, fileName);
@@ -138,8 +138,8 @@ public class FileMethods {
         return getFileTuple(path);
     }
 
-
     private FileTuple getProfile(Experiment e, String metaData) {
+
         for (FileTuple ft : e.getFiles()) {
             if (ft.type.equalsIgnoreCase("profile")
                     && ft.metaData.equals(metaData)) {
@@ -149,21 +149,21 @@ public class FileMethods {
         return null;
     }
 
-
     private String getParentFolder(String filePath) {
+
         int filenameIndex = filePath.lastIndexOf(File.separator);
         return filePath.substring(0, filenameIndex + 1);
     }
-    
+
     private String getFileName(String filePath) {
+
         int filenameIndex = filePath.lastIndexOf(File.separator);
         return filePath.substring(filenameIndex + 1);
     }
 
-
     /**
      * Returns the FileTuple object associated with the given filePath.
-     * 
+     *
      * @param filePath
      *            String
      * @return FileTuple - The corresponding FileTuple or null if no such file
@@ -173,7 +173,7 @@ public class FileMethods {
      */
     public FileTuple getFileTuple(String filePath) throws SQLException {
 
-        String query = "SELECT * FROM File WHERE Path ~~* ?";
+        String query = "SELECT * FROM File WHERE Path = ?";
         PreparedStatement stmt = conn.prepareStatement(query);
         stmt.setString(1, filePath);
         ResultSet rs = stmt.executeQuery();
@@ -186,10 +186,9 @@ public class FileMethods {
         return null;
     }
 
-
     /**
      * Returns the FileTuple object associated with the given fileID.
-     * 
+     *
      * @param fileID
      *            int
      * @return FileTuple - The corresponding FileTuple or null if no such file
@@ -212,21 +211,33 @@ public class FileMethods {
         return null;
     }
 
-
     /**
      * Deletes a file from the database and the disk. Should throw an
      * IOException if the method failed to delete the file from disk.
-     * 
+     *
      * @param path
      *            String - the path to the file.
      * @return int - the number of deleted tuples in the database.
      * @throws SQLException
      *             if the query does not succeed
+     * @throws IOException
      */
-    public int deleteFile(String path) throws SQLException {
+    public int deleteFile(String path) throws SQLException, IOException {
+
         File fileToDelete = new File(path);
         if (fileToDelete.exists()) {
             fileToDelete.delete();
+        }
+
+        FileTuple ft = getFileTuple(path);
+
+        if (ft == null) {
+            throw new IOException("Could not find file at path " + path);
+        }
+
+        File parentFolder = new File(ft.getParentFolder());
+        if (ft.type.equalsIgnoreCase("profile") && isEmptyFolder(parentFolder)) {
+            parentFolder.delete();
         }
 
         String statementStr = "DELETE FROM File " + "WHERE (Path ~~* ?)";
@@ -238,42 +249,76 @@ public class FileMethods {
         return resCount;
     }
 
-
     /**
      * Deletes a file from the database and the disk using the fileID. Should
      * throw an IOException if the method failed to delete the file from disk.
-     * 
+     *
      * @param fileID
      *            int - the fileID of the file to be deleted.
      * @return 1 if deletion was successful, else 0.
-     * @throws SQLException
+     * @throws SQLException If the database could not be contacted
+     * @throws IOException If the FileID does not exist in the database
      */
-    public int deleteFile(int fileID) throws SQLException {
+    public int deleteFile(int fileID) throws SQLException, IOException {
 
-        String query1 = "SELECT Path FROM File " + "WHERE FileID = ?";
-        String query2 = "DELETE FROM File " + "WHERE FileID = ?";
-        int res = 0;
+        FileTuple ft = getFileTuple(fileID);
 
-        PreparedStatement stmt = conn.prepareStatement(query1);
-        stmt.setInt(1, fileID);
-        ResultSet rs = stmt.executeQuery();
-        if (rs.next()) {
-            File fileToDelete = new File(rs.getString("Path"));
-            if (fileToDelete.exists()) {
-                fileToDelete.delete();
-            }
-            stmt = conn.prepareStatement(query2);
-            stmt.setInt(1, fileID);
-            res = stmt.executeUpdate();
-            stmt.close();
+        if (ft == null) {
+            throw new IOException("Could not find file with ID " + fileID);
         }
+
+        File fileToDelete = new File(ft.path);
+
+        if (fileToDelete.exists()) {
+
+            fileToDelete.delete();
+
+            File parentFolder = new File(ft.getParentFolder());
+            if (ft.type.equalsIgnoreCase("profile")
+            		&& isEmptyFolder(parentFolder)) {
+                parentFolder.delete();
+            }
+        }
+
+        String query = "DELETE FROM File " + "WHERE FileID = ?";
+        PreparedStatement stmt = conn.prepareStatement(query);
+        stmt.setInt(1, fileID);
+        int res = stmt.executeUpdate();
+        stmt.close();
         return res;
     }
 
+    private boolean isEmptyFolder(File f) {
+
+        if (f.exists() && f.listFiles() != null && f.listFiles().length == 0) {
+            return true;
+        }
+
+        return false;
+    }
+
+    /**
+     * Recursively deletes a folder with all it's subfolders and files.
+     * @param folder the folder to delete.
+     */
+    public void recursiveDelete(File folder) {
+
+        File[] contents = folder.listFiles();
+
+        if (contents == null || contents.length == 0) {
+            folder.delete();
+        } else {
+            for (File f : contents) {
+                recursiveDelete(f);
+            }
+        }
+
+        folder.delete();
+    }
 
     /**
      * Checks if the file with the specified fileID exists in the database.
-     * 
+     *
      * @param fileID
      *            int - the fileID of the file.
      * @return true if the file exists, else false.
@@ -297,6 +342,12 @@ public class FileMethods {
         return hasResult;
     }
 
+    /**
+     * Sets the status of a file to "Done".
+     * @param fileID the ID of the file to set to "Done".
+     * @return the number of tuples updated.
+     * @throws SQLException
+     */
 
     public int fileReadyForDownload(int fileID) throws SQLException {
 
@@ -313,12 +364,11 @@ public class FileMethods {
         return resCount;
     }
 
-
     /**
      * Changes the Filename for a specific file with given fileID. This method
      * affects bothe the saved file name, but also the entries path and fileName
      * in database.
-     * 
+     *
      * @return resCount int, the number of rows affected by the change.
      * @throws SQLException
      *             if failed to send query,
@@ -378,43 +428,46 @@ public class FileMethods {
         return resCount;
     }
 
-
     public FileTuple addGeneratedFile(String expId, int fileType,
             String filePath, String inputFileName, String metaData,
-            String uploader, boolean isPrivate, String grVersion) throws SQLException, IOException {
-        
+            String uploader, boolean isPrivate, String grVersion)
+            throws SQLException, IOException {
+
         Experiment e = expMethods.getExperiment(expId);
+
         if (e == null) {
             throw new IOException("The experiment " + expId
                     + " does not exist!");
         }
-        expId = e.getID();
 
+        expId = e.getID();
         String inputFilePath = null;
+
         if (inputFileName != null) {
             inputFilePath = getParentFolder(filePath) + inputFileName;
         }
 
         String query = "INSERT INTO File "
                 + "(Path, FileType, FileName, Date, MetaData, InputFilePath, "
-                + "Author, Uploader, IsPrivate, ExpID, GRVersion) "
-                + "VALUES (?, ?, ?, CURRENT_TIMESTAMP, ?, ?, 'Genomizer', ?, ?, ?, ?)";
+                + "Author, Uploader, IsPrivate, ExpID, GRVersion, Status) "
+                + "VALUES (?, ?, ?, CURRENT_TIMESTAMP, ?, ?, 'Genomizer'," +
+                " ?, ?, ?, ?, 'Done')";
         PreparedStatement stmt = conn.prepareStatement(query);
         stmt.setString(1, filePath);
 
         switch (fileType) {
-            case FileTuple.RAW:
-                stmt.setString(2, "Raw");
-                break;
-            case FileTuple.PROFILE:
-                stmt.setString(2, "Profile");
-                break;
-            case FileTuple.REGION:
-                stmt.setString(2, "Region");
-                break;
-            default:
-                stmt.setString(2, "Other");
-                break;
+        case FileTuple.RAW:
+            stmt.setString(2, "Raw");
+            break;
+        case FileTuple.PROFILE:
+            stmt.setString(2, "Profile");
+            break;
+        case FileTuple.REGION:
+            stmt.setString(2, "Region");
+            break;
+        default:
+            stmt.setString(2, "Other");
+            break;
         }
 
         stmt.setString(3, getFileName(filePath));
